nanshe/advanced_image_processing.py: Moved conversion out of for loop for extract_f0.

+++ nanshe/advanced_image_processing.py
@@ -214,9 +214,9 @@ def extract_f0(new_data,
     # TODO: Check what border treatment to use
     spatial_smoothing_gaussian_filter.setBorderTreatment(vigra.filters.BorderTreatmentMode.BORDER_TREATMENT_REFLECT)
 
-    new_data_spatially_smoothed = new_data_quantiled
+    new_data_spatially_smoothed = new_data_quantiled.astype(numpy.float32)
     for d in xrange(1, new_data_quantiled.ndim):
-        new_data_spatially_smoothed = vigra.filters.convolveOneDimension(new_data_spatially_smoothed.astype(numpy.float32),
+        new_data_spatially_smoothed = vigra.filters.convolveOneDimension(new_data_spatially_smoothed,
                                                                          d,
                                                                          spatial_smoothing_gaussian_filter)
